+++ src/main/java/com/hbng/miniapp/charger/services/CustomerService.java
@@ -2,6 +2,7 @@ package com.hbng.miniapp.charger.services;
 
 import com.hbng.miniapp.charger.model.Customer;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -9,12 +10,12 @@ public interface CustomerService {
 
      List<Customer> findAll();
 
-     List<Customer> findByAmount(double amount);
+     List<Customer> findByAmount(BigDecimal amount);
      Customer findById(Long theId);
 
      void save(Customer customer);
 
      void deleteById(Long theId);
 
-     Map<String, Integer> myMap(double amount);
+     Map<String, Integer> myMap(BigDecimal amount);
 }
